Moved data manipulation into reducer for convenience
