+++ src/js/helpers/__mocks__/weather.js
@@ -1,0 +1,4 @@
+
+export const groupForecastsByDay = (forecast) => {
+  return forecast;
+}
