+++ src/js/helpers/__tests__/weather-tests.js
@@ -1,0 +1,11 @@
+jest.unmock('../weather');
+
+// import { groupForecastsByDay } from '../weather';
+
+describe('weather helpers', () => {
+
+  it('todo', () => {
+    expect(1).toEqual(1);
+  });
+
+});
