+++ src/js/reducers/forecast.js
@@ -1,9 +1,11 @@
+import { groupForecastsByDay } from '../helpers/weather';
 
 const forecast = (state = {}, action) => {
   switch (action.type) {
     case 'NEW_FORECAST':
       const newState = Object.assign({}, state);
-      newState[action.location] = action.forecast;
+      const groupedForecasts = groupForecastsByDay(action.forecast);
+      newState[action.location] = groupedForecasts;
       return newState;
     default:
       return state
